move shadow yaml to shadow-common
